Refactor/domain database (#37)
* feat: Use UUID type for ID field with BINARY(16) storage

* refactor: Change String ID to UUID

* feat: Add custom JPQL query for fetching article summaries

* test: Fix UUID generation in test cases to ensure compatibility

* fix: CommentRepository 중복 제거

* refactor: spotless 적용

---------

+++ src/main/java/io/spring/api/article/response/ArticleSummaryResponse.java
@@ -9,8 +9,8 @@ public record ArticleSummaryResponse(
     String title,
     LocalDateTime createdAt,
     UserResponse author,
-    int tagCount,
-    int commentCount) {
+    long tagCount,
+    long commentCount) {
   public static ArticleSummaryResponse from(Article article) {
     return new ArticleSummaryResponse(
         article.getSlug(),
